DataRequestor/requirements.txt
python-snap7==1.3
requests==2.31.0
paho-mqtt